Conflicts:
	inc/INIT.php
	inc/config.ini.sample
	lib/Controller/convertFileController.php
	lib/Utils/FileFormatConverter.php
	public/js/lib/fileupload/main.js

+++ lib/Model/FilesStorage.php
@@ -30,19 +30,29 @@ class FilesStorage {
 
     private $filesDir;
     private $cacheDir;
+    private $zipDir;
 
-    public function __construct( $files = false, $cache = false ) {
+    public function __construct( $files = null, $cache = null, $zip = null ) {
+
         //override default config
         if ( $files ) {
             $this->filesDir = $files;
         } else {
             $this->filesDir = INIT::$FILES_REPOSITORY;
         }
+
         if ( $cache ) {
             $this->cacheDir = $cache;
         } else {
             $this->cacheDir = INIT::$CACHE_REPOSITORY;
         }
+
+        if( $zip ){
+            $this->zipDir = $zip;
+        } else {
+            $this->zipDir = INIT::$ZIP_REPOSITORY;
+        }
+
     }
 
     /**
@@ -139,9 +149,11 @@ class FilesStorage {
         $cacheTree = implode( DIRECTORY_SEPARATOR, $this->_composeCachePath( $hash ) );
 
         //ensure old stuff is overwritten
-        if ( is_dir( $this->cacheDir . DIRECTORY_SEPARATOR . $cacheTree . "|" . $lang ) ) {
-            Utils::deleteDir( $this->cacheDir . DIRECTORY_SEPARATOR . $cacheTree . "|" . $lang );
-        }
+        // ?? why ?? if we have more than once the same file we are not
+        // be benefiting from the advantages of hard links
+//        if ( is_dir( $this->cacheDir . DIRECTORY_SEPARATOR . $cacheTree . "|" . $lang ) ) {
+//            Utils::deleteDir( $this->cacheDir . DIRECTORY_SEPARATOR . $cacheTree . "|" . $lang );
+//        }
 
         //create cache dir structure
         mkdir( $this->cacheDir . DIRECTORY_SEPARATOR . $cacheTree . "|" . $lang, 0755, true );
@@ -153,16 +165,17 @@ class FilesStorage {
         //if it's not an xliff as original
         if ( !$originalPath ) {
 
-            //set original moval as successful
-            $outcome1 = true;
-
+            //if there is not an original path this is an unconverted file,
+            // the original does not exists
+            // detect which type of xliff
+            //check also for the extension, if already present do not force
             $fileType = DetectProprietaryXliff::getInfo( $xliffPath );
-            if ( !$fileType[ 'proprietary' ] ) {
+            if ( !$fileType[ 'proprietary' ] && $fileType[ 'info' ][ 'extension' ] != 'sdlxliff' ) {
                 $force_extension = '.sdlxliff';
             }
 
             //use original xliff
-            $xliffDestination = $cacheDir . DIRECTORY_SEPARATOR . "work" . DIRECTORY_SEPARATOR . basename( $xliffPath ) . @$force_extension;
+            $xliffDestination = $cacheDir . DIRECTORY_SEPARATOR . "work" . DIRECTORY_SEPARATOR . FilesStorage::basename_fix( $xliffPath ) . @$force_extension;
         } else {
             //move original
             $raw_file_path = explode(DIRECTORY_SEPARATOR, $originalPath);
@@ -182,7 +195,7 @@ class FilesStorage {
             if ( file_exists( $manifestFile ) ) {
                 Log::doLog( "Alternative Conversion detected" );
                 $file_extension = '.xlf';
-            } else{
+            } else {
                 Log::doLog( "Normal Conversion detected" );
                 $file_extension = '.sdlxliff';
             }
@@ -195,25 +208,139 @@ class FilesStorage {
         //In Unix you can't rename or move between filesystems,
         //Instead you must copy the file from one source location to the destination location, then delete the source.
         $outcome2 = copy( $xliffPath, $xliffDestination );
-        unlink( $xliffPath );
 
         if ( !$outcome2 ) {
             //Original directory deleted!!!
-            //CLEAR ALL CACHE
+            //CLEAR ALL CACHE - FATAL
             Utils::deleteDir( $this->cacheDir . DIRECTORY_SEPARATOR . $cacheTree . "|" . $lang );
             return $outcome2;
         }
+
+        unlink( $xliffPath );
+        return true;
+
+    }
+
+    /**
+     * Make a temporary cache copy for the original zip file
+     *
+     * @param $hash
+     * @param $zipPath
+     *
+     * @return bool
+     */
+    public function cacheZipArchive( $hash, $zipPath ){
+
+        $thisZipDir = $this->zipDir . DIRECTORY_SEPARATOR . $hash . "__##originalZip##";
+
+        //ensure old stuff is overwritten
+        if ( is_dir( $thisZipDir ) ) {
+            Utils::deleteDir( $thisZipDir );
+        }
+
+        //create cache dir structure
+        $created = mkdir( $thisZipDir, 0755, true );
+
+        if( !$created ){
+            return $created;
+        }
+
+        //move original
+        $outcome1 = copy( $zipPath, $thisZipDir . DIRECTORY_SEPARATOR . FilesStorage::basename_fix( $zipPath ) );
+
+        if( !$outcome1 ){
+            //Original directory deleted!!!
+            //CLEAR ALL CACHE
+            Utils::deleteDir( $this->zipDir . DIRECTORY_SEPARATOR . $hash . "__##originalZip##" );
+            return $outcome1;
+        }
+
+        unlink( $zipPath );
+
+        //link this zip to the upload directory by creating a file name as the ash of the zip file
+        touch( dirname( $zipPath ) . DIRECTORY_SEPARATOR . $hash . "__##originalZip##" );
 
         return true;
 
     }
 
-    public function linkSessionToCache( $hash, $lang, $uid ) {
+    public function linkZipToProject( $create_date, $zipHash, $projectID ){
+
+        $datePath = date_create( $create_date )->format( 'Ymd' );
+
+        $fileName = FilesStorage::basename_fix( $this->getSingleFileInPath( $this->zipDir . DIRECTORY_SEPARATOR . $zipHash ) );
+
+        //destination dir
+        $newZipDir  = $this->zipDir . DIRECTORY_SEPARATOR . $datePath . DIRECTORY_SEPARATOR . $projectID ;
+
+        //check if doesn't exist
+        if ( !is_dir( $newZipDir ) ) {
+            //make files' directory structure
+            if( ! mkdir( $newZipDir, 0755, true ) ) {
+                return false;
+            }
+        }
+
+        //link original
+        $outcome1 = link( $this->getSingleFileInPath( $this->zipDir . DIRECTORY_SEPARATOR . $zipHash ) , $newZipDir . DIRECTORY_SEPARATOR . $fileName );
+
+        if( !$outcome1 ){
+            //Failed to copy the original file zip
+            return $outcome1;
+        }
+
+        Utils::deleteDir( $this->zipDir . DIRECTORY_SEPARATOR . $zipHash );
+
+        return true;
+
+    }
+
+    public function getOriginalZipPath( $projectDate, $projectID, $zipName ){
+
+        $datePath = date_create( $projectDate )->format('Ymd');
+        $zipDir = $this->zipDir . DIRECTORY_SEPARATOR . $datePath . DIRECTORY_SEPARATOR . $projectID . DIRECTORY_SEPARATOR . $zipName;
+
+        return $zipDir;
+
+    }
+
+    public function getHashesFromDir( $dirToScan ){
+
+        //fetch cache links, created by converter, from a directory
+        $linkFiles = scandir( $dirToScan );
+        $zipFilesHash = array();
+        $filesHashInfo    = array();
+        //remove dir hardlinks, as uninteresting, as well as regular files; only hash-links
+        foreach ( $linkFiles as $k => $linkFile ) {
+
+            if( strpos( $linkFile, "__##originalZip##" ) !== false ){
+                $zipFilesHash[ ] = $linkFile;
+                unset( $linkFiles[ $k ] );
+            } elseif ( strpos( $linkFile, '.' ) !== false or strpos( $linkFile, '|' ) === false ) {
+                unset( $linkFiles[ $k ] );
+            } else {
+                $filesHashInfo[ 'sha' ][] = $linkFiles[ $k ];
+                $filesHashInfo[ 'fileName' ][ $linkFiles[ $k ] ] = file(
+                        $dirToScan . DIRECTORY_SEPARATOR . $linkFiles[ $k ],
+                        FILE_IGNORE_NEW_LINES | FILE_SKIP_EMPTY_LINES
+                );
+            }
+
+        }
+
+        return array(
+                'conversionHashes' => $filesHashInfo,
+                'zipHashes'        => $zipFilesHash
+        );
+
+    }
+
+    public function linkSessionToCache( $hash, $lang, $uid, $realFileName ) {
         //get upload dir
         $dir = INIT::$UPLOAD_REPOSITORY . DIRECTORY_SEPARATOR . $uid;
 
         //create a file in it, named after cache position on storage
-        return touch( $dir . DIRECTORY_SEPARATOR . $hash . "|" . $lang );
+        return file_put_contents( $dir . DIRECTORY_SEPARATOR . $hash . "|" . $lang, $realFileName . "\n" , FILE_APPEND | LOCK_EX );
     }
 
     /*
@@ -229,7 +356,12 @@ class FilesStorage {
             if ( $fileInfo->isDot() || $fileInfo->isDir() ) {
                 continue;
             }
-            if ( stripos( $fileInfo->getFilename(), $shasum ) !== false ) {
+
+            //remove only the wrong languages, the same code|language must be
+            // retained because of the file name append
+            if ( $fileInfo->getFilename() != $linkFile &&
+                    stripos( $fileInfo->getFilename(), $shasum ) !== false ) {
+
                 unlink( $fileInfo->getPathname() );
                 Log::doLog( "Deleted Hash " . $fileInfo->getPathname() );
             }
@@ -237,7 +369,7 @@ class FilesStorage {
 
     }
 
-    public function moveFromCacheToFileDir( $dateHashPath, $lang, $idFile ) {
+    public function moveFromCacheToFileDir( $dateHashPath, $lang, $idFile, $newFileName = null ) {
 
         list( $datePath, $hash ) = explode( DIRECTORY_SEPARATOR, $dateHashPath );
         $cacheTree = implode( DIRECTORY_SEPARATOR, $this->_composeCachePath( $hash ) );
@@ -260,24 +392,41 @@ class FilesStorage {
 
         //make links from cache to files
         //BUG: this stuff may not work if FILES and CACHES are on different filesystems
-        //FIX: we could check in advance and, in case, use copy instead of links
-
         //orig, suppress error because of xliff files have not original one
-        $filePath = $this->getSingleFileInPath( $cacheDir . DIRECTORY_SEPARATOR . "orig" );
-        if( is_file( $filePath ) ){
-            $res &= link( $filePath, $fileDir . DIRECTORY_SEPARATOR . "orig" . DIRECTORY_SEPARATOR . basename( $filePath ) );
+        $origFilePath = $this->getSingleFileInPath( $cacheDir . DIRECTORY_SEPARATOR . "orig" );
+        $tmpOrigFileName = $origFilePath;
+        if( is_file( $origFilePath ) ){
+
+            /*
+             * Force the new filename if it is provided
+             */
+            if ( !empty( $newFileName ) ){
+                $tmpOrigFileName = $newFileName;
+            }
+            $res &= link( $origFilePath, $fileDir . DIRECTORY_SEPARATOR . "orig" . DIRECTORY_SEPARATOR . FilesStorage::basename_fix( $tmpOrigFileName ) );
+
         }
 
         //work
-        $filePath = $this->getSingleFileInPath( $cacheDir . DIRECTORY_SEPARATOR . "work" );
-        $res &= link( $filePath, $fileDir . DIRECTORY_SEPARATOR . "xliff" . DIRECTORY_SEPARATOR . basename( $filePath ) );
+        /*
+         * Force the new filename if it is provided
+         */
+        $convertedFilePath = $this->getSingleFileInPath( $cacheDir . DIRECTORY_SEPARATOR . "work" );
+        $tmpConvertedFilePath = $convertedFilePath;
+        if ( !empty( $newFileName ) ){
+            if( !DetectProprietaryXliff::isXliffExtension( FilesStorage::pathinfo_fix( $newFileName ) ) ){
+                $convertedExtension = FilesStorage::pathinfo_fix( $convertedFilePath, PATHINFO_EXTENSION );
+                $tmpConvertedFilePath = $newFileName . "." . $convertedExtension;
+            }
+        }
+        $res &= link( $convertedFilePath, $fileDir . DIRECTORY_SEPARATOR . "xliff" . DIRECTORY_SEPARATOR . FilesStorage::basename_fix( $tmpConvertedFilePath ) );
 
         //check if manifest from a LongHorn conversion exists
         $manifestFile = $cacheDir . DIRECTORY_SEPARATOR . "manifest.rkm";
         if ( file_exists( $manifestFile ) ) {
-            $res &= link( $manifestFile, $fileDir . DIRECTORY_SEPARATOR . "package" . DIRECTORY_SEPARATOR . basename( $manifestFile ) );
-            $res &= link( $filePath, $fileDir . DIRECTORY_SEPARATOR . "package" . DIRECTORY_SEPARATOR . "orig" . DIRECTORY_SEPARATOR . basename( $filePath ) );
-            $res &= link( $filePath, $fileDir . DIRECTORY_SEPARATOR . "package" . DIRECTORY_SEPARATOR . "work" . DIRECTORY_SEPARATOR . basename( $filePath ) );
+            $res &= link( $manifestFile, $fileDir . DIRECTORY_SEPARATOR . "package" . DIRECTORY_SEPARATOR . FilesStorage::basename_fix( $manifestFile ) );
+            $res &= link( $origFilePath, $fileDir . DIRECTORY_SEPARATOR . "package" . DIRECTORY_SEPARATOR . "orig" . DIRECTORY_SEPARATOR . FilesStorage::basename_fix( $tmpOrigFileName ) );
+            $res &= link( $convertedFilePath, $fileDir . DIRECTORY_SEPARATOR . "package" . DIRECTORY_SEPARATOR . "work" . DIRECTORY_SEPARATOR . FilesStorage::basename_fix( $tmpConvertedFilePath ) );
         }
 
         if( !$res ){
@@ -331,7 +480,7 @@ class FilesStorage {
         if ( !empty( $id_file ) ) {
             $where_id_file = " and fj.id_file=$id_file";
         }
-        $query = "select fj.id_file, f.filename, f.id_project, j.source, mime_type, sha1_original_file from files_job fj
+        $query = "select fj.id_file, f.filename, f.id_project, j.source, mime_type, sha1_original_file, create_date from files_job fj
 			inner join files f on f.id=fj.id_file
 			inner join jobs j on j.id=fj.id_job
 			where fj.id_job=$id_job $where_id_file and j.password='$password'";
@@ -404,7 +553,6 @@ class FilesStorage {
         return $results;
     }
 
-    //TODO FIX
     /**
      * Backwards compatibility method and forward
      *
@@ -515,6 +663,61 @@ class FilesStorage {
         $results = $db->fetch_array( $query );
 
         return $results[ 0 ][ 'xliff_file' ];
+    }
+
+    /**
+     * @param $path
+     *
+     * @return mixed
+     */
+    public static function basename_fix( $path ) {
+        $rawPath  = explode( DIRECTORY_SEPARATOR, $path );
+        $basename = array_pop( $rawPath );
+
+        return $basename;
+    }
+
+    /**
+     * @param     $path
+     * @param int $options
+     *
+     * @return array|mixed
+     */
+    public static function pathinfo_fix( $path, $options=15) {
+        $rawPath = explode( DIRECTORY_SEPARATOR, $path );
+
+        $basename = array_pop( $rawPath );
+        $dirname  = implode( DIRECTORY_SEPARATOR, $rawPath );
+
+        $explodedFileName = explode( ".", $basename );
+        $extension        = strtolower( array_pop( $explodedFileName ) );
+        $filename         = implode( ".", $explodedFileName );
+
+        $return_array = array();
+
+        $flagMap = array(
+                'dirname'   => PATHINFO_DIRNAME,
+                'basename'  => PATHINFO_BASENAME,
+                'extension' => PATHINFO_EXTENSION,
+                'filename'  => PATHINFO_FILENAME
+        );
+
+        // foreach flag, add in $return_array the corresponding field,
+        // obtained by variable name correspondance
+        foreach ( $flagMap as $field => $i ) {
+            //binary AND
+            if ( ( $options & $i ) > 0 ) {
+                //variable substitution: $field can be one between 'dirname', 'basename', 'extension', 'filename'
+                // $$field gets the value of the variable named $field
+                $return_array[ $field ] = $$field;
+            }
+        }
+
+        if ( count( $return_array ) == 1 ) {
+            $return_array = array_pop( $return_array );
+        }
+
+        return $return_array;
     }
 }
